Further commits to make this a relatively usable interface tool.
